✂️ Split code in multiple modules
