+++ src/checks/mod.rs
@@ -1,0 +1,4 @@
+pub mod author_connected_to_voice;
+pub mod bob_has_category;
+pub mod sent_in_bob;
+pub mod preset_exists;
