+++ src/commands/bob/mod.rs
@@ -1,0 +1,13 @@
+pub mod build;
+pub mod save;
+pub mod load;
+
+use serenity::framework::standard::macros::*;
+
+use crate::commands::bob::build::BUILD_COMMAND;
+use crate::commands::bob::save::SAVE_COMMAND;
+use crate::commands::bob::load::LOAD_COMMAND;
+
+#[group]
+#[commands(build, save, load)]
+pub struct Bob;
